# Simplify Tracer.shouldObserve