--- a/changelog/@unreleased/pr-870.v2.yml
+++ b/changelog/@unreleased/pr-870.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Simplify Tracer.shouldObserve to avoid JIT friction
+  links:
+  - https://github.com/palantir/tracing-java/pull/870