@svc-autorelease 2022-03-09
Released [6.11.0](https://github.com/palantir/tracing-java/releases/tag/6.11.0)